Fix: size_t unsigned int 변경

+++ cpp07/ex01/incs/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: jihoh <jihoh@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/21 17:44:13 by jihoh             #+#    #+#             */
-/*   Updated: 2022/07/02 13:40:16 by jihoh            ###   ########.fr       */
+/*   Updated: 2022/07/03 15:13:27 by jihoh            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 # define ITER_CLASS_H
 
 template <typename T>
-void iter(T *array, size_t arraySize, void (*f)(const T &))
+void iter(T *array, unsigned int arraySize, void (*f)(const T &))
 {
-	for (size_t i = 0; i < arraySize; i++)
+	for (unsigned int i = 0; i < arraySize; i++)
 		f(array[i]);
 }
